Hardcode the basic Jupyter messages.
This is just following the spec at
http://jupyter-client.readthedocs.io/en/stable/messaging.html

We need a special constant, the message protocol version.  It’s a
required part of the message spec, even though my implementation doesn’t
depend on it yet.

We also need a UUID function.  Use `org-id-uuid`, because … Emacs
doesn’t provide a UUID builtin, I don’t want to make this depend on a
UUID library, *and* I don’t want to roll my own.  If it’s good enough
for Org, it’s good enough for me.