BUG FIX read all IOPub replies on distinct sockets
This enables us to send multiple requests to the kernel and reliably
separate out the various replies.  As of this writing I’m only seeing a
benefit to the session initialization, when we send two
kernel-info-requests in a row.

However I expect this will have other benefits down the road, maybe?
Like what about if we try to org babel eval an entire buffer?  Won’t all
the execute requests fire basically at once?