NOT QUITE BUG FIX?  wait a while for subscriptions to propagate
Previously, we waited 10 msec.  On my laptop, this isn’t enough time for
connections to settle before completion requests come back.  So bump it
up to waiting 100 msec.

Note that this is the worst possible way to fix a bug.  I have a TODO to
fix it the right way.