Use :dataframe header arg, instead of overloading :table.
The table overload is starting to mess with me, because what’s the
expected behavior if we return, say, a Python list that should be
formatted as a table but is not a dataframe?  My dataframe code is a
little too rigid, can’t handle that case.  So just don’t bother.