BUG FIX toggle python shell buffer name, don’t just set it.
I noticed this problem when I was working on a script, happily chugging
away with a dedicated inferior python repl, and then thought to throw
the code at an existing jupyter session for the hell of it.  When I
turned off jupyter-mode in the original script buffer … oops, I had
clobbered the awareness of the original repl.  Fixed!