BUG FIX actually pull down completions on the dot.
Holy cow!  So, turns out, checking the cached prefix for triviality is
wrong wrong wrong.  We should’ve been checking the regex match the whole
time!

I forget exactly how this wrong code came to be, and why it kinda sorta
worked before.  But it was definitely wrong, and changing the test makes
it work now.