BUG FIX accept that jupyter is sending back completions dupes.
Looks like it’s an experimental thing with the latest kernel.  They’re
moving towards sending back those jedi annotations, but for some reason
they’re pulling both the annotated completion *and* a copy of the same
completion with an “unknown” annotation.

Really annoying to me, at least, because this experimental stuff is
happening in the query reply metadata (good, I can ignore that)
but *also* affecting the behavior of the existing completion
candidates (ugh).  Oh well.  Fixed.