Automatically include a comma after json encoding alists.
The Org Export parser concatenates parsed results as it goes along, so
we must find a way to insert commas in between the JSON cells that are
the result of parsing.  It just winds up being easier to include a comma
in every parse result, then deal with that annoying final comma at the end.